# docs/api/openapi-patudos-scalar.yaml
openapi: 3.0.3
info:
  title: Patudos & Companhia API
  version: 1.1.0
  description: Documentação OpenAPI da API REST do backend do sistema Patudos & Companhia, preparada para visualização no
    Scalar. A API suporta autenticação JWT, gestão de utilizadores, animais, reservas, serviços, pagamentos e relatórios operacionais.
    Esta versão inclui a gestão de indisponibilidades de serviços por data, eliminação completa de serviços e atualização
    dos estados reais das reservas.
  contact:
    name: Equipa LI4 - Patudos & Companhia
servers:
  - url: http://localhost:8081
    description: Servidor local Spring Boot
...
security:
  - bearerAuth: []
paths:
...
  /api/servicos/disponiveis:
    get:
      tags:
        - Serviços
      summary: Listar serviços disponíveis
      operationId: listarServicosDisponiveis
      description: Lista o catálogo público de serviços disponíveis. Quando o parâmetro data é indicado, exclui também os
        serviços bloqueados nessa data específica.
      security: []
      responses:
        '200':
          description: Lista de serviços disponíveis globalmente e, se aplicável, na data indicada.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServicoResponse'
        '500':
          description: Erro interno inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
      parameters:
        - name: data
          in: query
          required: false
          schema:
            type: string
            format: date
            example: '2026-05-25'
          description: Data opcional para filtrar os serviços disponíveis nesse dia.
  /api/servicos:
    get:
      tags:
        - Serviços
      summary: Listar todos os serviços
      operationId: listarServicos
      description: Lista todo o catálogo de serviços.
      responses:
        '200':
          description: Lista de serviços.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServicoResponse'
        '401':
          description: Não autenticado. Token JWT ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500':
          description: Erro interno inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
    post:
      tags:
        - Serviços
      summary: Criar serviço
      operationId: criarServico
      description: Cria um serviço no catálogo. Apenas DIRECAO ou ADMIN.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicoRequest'
      responses:
        '201':
          description: Serviço criado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicoResponse'
        '400':
          description: Pedido inválido ou regra de negócio violada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '401':
          description: Não autenticado. Token JWT ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '403':
          description: Acesso recusado. O utilizador autenticado não tem permissões suficientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500':
          description: Erro interno inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
  /api/servicos/{id}:
    get:
      tags:
        - Serviços
      summary: Obter serviço por ID
      operationId: obterServico
      description: Consulta um serviço do catálogo.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador numérico do recurso.
      responses:
        '200':
          description: Serviço encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicoResponse'
        '401':
          description: Não autenticado. Token JWT ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '404':
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500':
          description: Erro interno inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
    put:
      tags:
        - Serviços
      summary: Editar serviço
      operationId: editarServico
      description: Atualiza um serviço do catálogo. Apenas DIRECAO ou ADMIN.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador numérico do recurso.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicoRequest'
      responses:
        '200':
          description: Serviço atualizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicoResponse'
        '400':
          description: Pedido inválido ou regra de negócio violada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '401':
          description: Não autenticado. Token JWT ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '403':
          description: Acesso recusado. O utilizador autenticado não tem permissões suficientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '404':
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500':
          description: Erro interno inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
    delete:
      tags:
        - Serviços
      summary: Eliminar serviço do catálogo
      operationId: eliminarServico
      description: Remove completamente um serviço do catálogo. Apenas DIRECAO ou ADMIN. Difere de PATCH /api/servicos/{id}/toggle,
        que apenas ativa/desativa o serviço globalmente.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador do serviço a eliminar.
      responses:
        '204':
          description: Serviço eliminado com sucesso.
        '401': &id001
          description: Não autenticado. Token JWT ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '403': &id002
          description: Acesso recusado. O utilizador autenticado não tem permissões suficientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '404': &id003
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500': &id004
          description: Erro interno inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
  /api/servicos/{id}/toggle:
    patch:
      tags:
        - Serviços
      summary: Alternar disponibilidade do serviço
      operationId: toggleDisponibilidadeServico
      description: Ativa ou desativa um serviço do catálogo. Apenas DIRECAO ou ADMIN.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador numérico do recurso.
      responses:
        '204':
          description: Disponibilidade alterada.
        '401':
          description: Não autenticado. Token JWT ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '403':
          description: Acesso recusado. O utilizador autenticado não tem permissões suficientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '404':
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500':
          description: Erro interno inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
  /api/servicos/reserva/{reservaId}:
    post:
      tags:
        - Serviços
      summary: Associar serviço a reserva
      operationId: associarServicoAReserva
      description: Associa um serviço existente a uma reserva.
      parameters:
        - name: reservaId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador da reserva.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssociarServicoRequest'
      responses:
        '201':
          description: Serviço associado à reserva.
        '400':
          description: Pedido inválido ou regra de negócio violada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '401':
          description: Não autenticado. Token JWT ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '403':
          description: Acesso recusado. O utilizador autenticado não tem permissões suficientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '404':
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500':
          description: Erro interno inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
  /api/servicos/reserva/{reservaId}/{reservaServicoId}:
    delete:
      tags:
        - Serviços
      summary: Remover serviço de reserva
      operationId: removerServicoDeReserva
      description: Remove uma associação de serviço de uma reserva.
      parameters:
        - name: reservaId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador da reserva.
        - name: reservaServicoId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador da associação reserva-serviço.
      responses:
        '204':
          description: Serviço removido da reserva.
        '401':
          description: Não autenticado. Token JWT ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '403':
          description: Acesso recusado. O utilizador autenticado não tem permissões suficientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '404':
          description: Recurso não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500':
          description: Erro interno inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
  /api/servicos/dia:
    get:
      tags:
        - Serviços
      summary: Listar serviços do dia
      operationId: listarServicosDoDia
      description: Lista serviços a executar numa data. Se a data for omitida, usa a data atual do servidor.
      parameters:
        - name: data
          in: query
          required: false
          schema:
            type: string
            format: date
          description: Data de execução.
      responses:
        '200':
          description: Lista de serviços agendados para o dia indicado.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServicoAgendadoResponse'
        '401':
          description: Não autenticado. Token JWT ausente, inválido ou expirado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '403':
          description: Acesso recusado. O utilizador autenticado não tem permissões suficientes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '500':
          description: Erro interno inesperado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
...
  /api/servicos/{id}/indisponibilidades:
    post:
      tags:
        - Serviços
      summary: Bloquear serviço numa data específica
      operationId: bloquearServicoNumaData
      description: Regista uma indisponibilidade pontual de um serviço numa data específica, sem desativar o serviço globalmente.
        Apenas DIRECAO ou ADMIN.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador do serviço a bloquear.
        - name: data
          in: query
          required: true
          schema:
            type: string
            format: date
            example: '2026-05-25'
          description: Data em que o serviço fica indisponível.
      responses:
        '201':
          description: Indisponibilidade criada com sucesso.
        '400': &id005
          description: Pedido inválido ou regra de negócio violada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErroResponse'
        '401': *id001
        '403': *id002
        '404': *id003
        '500': *id004
    delete:
      tags:
        - Serviços
      summary: Remover bloqueio de serviço numa data específica
      operationId: desbloquearServicoNumaData
      description: Remove uma indisponibilidade pontual previamente registada, voltando a disponibilizar o serviço nessa data.
        Apenas DIRECAO ou ADMIN.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador do serviço a desbloquear.
        - name: data
          in: query
          required: true
          schema:
            type: string
            format: date
            example: '2026-05-25'
          description: Data cuja indisponibilidade deve ser removida.
      responses:
        '204':
          description: Indisponibilidade removida com sucesso.
        '401': *id001
        '403': *id002
        '404': *id003
        '500': *id004
  /api/servicos/reserva/{reservaId}/{reservaServicoId}/realizado:
    patch:
      tags:
        - Serviços
      summary: Marcar serviço agendado como realizado
      operationId: marcarServicoAgendadoComoRealizado
      description: Marca um serviço associado a uma reserva como realizado. Apenas FUNC_OPERACIONAL ou ADMIN.
      parameters:
        - name: reservaId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador da reserva.
        - name: reservaServicoId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Identificador da associação reserva-serviço.
      responses:
        '200':
          description: Serviço marcado como realizado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicoAgendadoResponse'
        '400': *id005
        '401': *id001
        '403': *id002
        '404': *id003
        '500': *id004
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Usar o token devolvido por POST /api/auth/login no formato: Bearer <token>.'
  schemas:
...
    EstadoReserva:
      type: string
      enum:
        - PENDENTE
        - EM_ESTADIA
        - CONCLUIDA
        - CANCELADA
...
    ServicoRequest:
      type: object
      required:
        - nome
        - preco
        - capacidadeDiaria
      properties:
        nome:
          type: string
          example: Banho
        descricao:
          type: string
          example: Banho completo durante a estadia.
        preco:
          type: number
          format: double
          example: 20.0
        capacidadeDiaria:
          type: integer
          example: 5
    ServicoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 5
        nome:
          type: string
          example: Banho
        descricao:
          type: string
          example: Banho completo durante a estadia.
        preco:
          type: number
          format: double
          example: 20.0
        capacidadeDiaria:
          type: integer
          example: 5
        disponivel:
          type: boolean
          example: true
        datasIndisponiveis:
          type: array
          description: Datas em que o serviço está bloqueado/indisponível, apesar de poder continuar disponível globalmente.
          items:
            type: string
            format: date
          example:
            - '2026-05-25'
            - '2026-05-26'
    AssociarServicoRequest:
      type: object
      required:
        - servicoId
        - dataExecucao
      properties:
        servicoId:
          type: integer
          format: int64
          example: 5
        dataExecucao:
          type: string
          format: date
          example: '2026-06-03'
          description: Data em que o serviço deve ser executado. Deve estar dentro do período da reserva e não pode estar
            bloqueada para esse serviço.
...
    ErroResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        erro:
          type: string
          example: Bad Request
        mensagem:
          type: string
          example: Datas inválidas.
    ServicoAgendadoResponse:
      type: object
      properties:
        reservaServicoId:
          type: integer
          format: int64
          example: 301
        reservaId:
          type: integer
          format: int64
          example: 100
        servicoId:
          type: integer
          format: int64
          example: 5
        servicoNome:
          type: string
          example: Banho
        animalId:
          type: integer
          format: int64
          example: 10
        animalNome:
          type: string
          example: Tobias
        proprietarioNome:
          type: string
          example: Rui Abreu
        codigoEspaco:
          type: string
          nullable: true
          example: CA1
        dataExecucao:
          type: string
          format: date
          example: '2026-06-03'
        realizado:
          type: boolean
          example: false
        preco:
          type: number
          format: double
          example: 20.0
        estadoReserva:
          $ref: '#/components/schemas/EstadoReserva'